move prometheus to sep file and add tests
